refactor: lms.js -> dataApiClient.js
