+++ components/role-list.js
@@ -4,7 +4,7 @@ import '@brightspace-ui/core/components/inputs/input-checkbox';
 import { css, html, LitElement } from 'lit-element';
 import { classMap } from 'lit-html/directives/class-map.js';
 import { fetchRoles as fetchDemoRoles } from '../model/fake-lms';
-import { fetchRoles } from '../model/lms';
+import { fetchRoles } from '../model/dataApiClient';
 import { heading3Styles } from '@brightspace-ui/core/components/typography/styles.js';
 import { Localizer } from '../locales/localizer';
