+++ engagement-dashboard.js
@@ -32,7 +32,7 @@ import { CurrentFinalGradesFilter } from './components/current-final-grade-card'
 import { Data } from './model/data.js';
 import { DiscussionActivityFilter } from './components/discussion-activity-card';
 import { ExportData } from './model/exportData';
-import { fetchData } from './model/lms.js';
+import { fetchData } from './model/dataApiClient.js';
 import { fetchData as fetchDemoData } from './model/fake-lms.js';
 import { FilteredData } from './model/filteredData';
 import { heading3Styles } from '@brightspace-ui/core/components/typography/styles';
